+++ inplaceeditform/fields.py
@@ -589,6 +589,8 @@ class AdaptorFileField(BaseAdaptorField):
         if not value:
             return ''
         config = deepcopy(self.config)
+        if self.request.method == 'POST' and config.get('msie', None):
+            config['innerHtml'] = ''
         context = {'value': value,
                    'obj': self.obj}
         config.update(context)
@@ -611,4 +613,7 @@ class AdaptorImageField(AdaptorFileField):
         return super(AdaptorImageField, self).render_media_field(template_name, extra_context=extra_context)
 
     def render_value(self, field_name=None, template_name='inplaceeditform/adaptor_image/render_value.html'):
+        if self.request.method == 'POST' and self.config.get('msie', None):
+            self.config['title'] = ""
+            self.config['alt'] = ""
         return super(AdaptorImageField, self).render_value(field_name=field_name, template_name=template_name)