+++ CHANGES.rst
@@ -10,6 +10,7 @@ Releases
 * Create a doc in readthedocs
 * Add unit test and integrate with `travis <https://travis-ci.org/Yaco-Sistemas/django-inplaceedit>`_
 * The django-inplaceedit now is less intrusive. **Attention** if you have customize django-inplaceedit, see this `changeset  <https://github.com/Yaco-Sistemas/django-inplaceedit/commit/c5cfdcce190b4fa8166b7500db711400baa9ea86>`_
+* Fix some errors with IE browser
 * Improvement in testing project
 * Move the settings to the module
 * Detail of Python3 compatibility